Added dropdown menu
We make the dropdown appear dynamically by using the `show` class. Bootstrap does this automatically if we import the bootstrap javascript files, but we do not want to do that in our Angular application. This is because Bootstrap.js is dependent on jQuery, which is not something that we want inside of an Angular application.

To avoid having to implement our own directives to show/hide bootstrap dropdowns and other components, we use `NgBootstrap` which implements these for us.

You can browse directives at https://ng-bootstrap.github.io