Minor fixes
Because we set href to #, clicking the dropdown will sometimes launch us to the front page before we have had a change to select an item. Delete the `href` class from the anchor.

We also fix so that the cursor turns into a pointer to make it more clear that the dropdown menu is clickable.

The Home button is also removed because we can always go back to the homepage by pressing the "website logo".

We also fixed the "My Orders" link not working.