Redirect user after logging in
When using Google login, the returnUrl query is lost. We must therefore save it in local storage to use it once we have completed the login process. Note that the code in AppComponent is a workaround because .else approach does not work for some reason. Because we are subscribing without implementing onDestroy, we technically have a memory leak, but because the AppComponent is the root component, it does not really make a difference.