Storing users in database
Our application has no concept of registering a user, so we save the user to the database every time they log in.

Ran into an annoying issue where the firebase does not find the database unless it ends with firebaseio.com. Because I had chosen a European server, the suffix was different and simply would not work at all. I had to recreate a new Firebase project from scratch and select a US server to resolve this issue.