Flattening shopping cart items
Our current solution did not allow us to access the quantity easily from the shopping cart. This is because we put all items related to the product under in its own object for every shopping cart item. Now, we have put only the necessary items needed to render the shopping cart page on the same level as the quantity, simplifying the structure in the database.